requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
bleach==6.3.0
boto3==1.42.76
botocore==1.42.76
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==3.1.2
Django==6.0.3
django-admin-interface==0.32.0
django-anymail==14.0
django-appconf==1.2.0
django-celery-beat==2.9.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-filter==25.2
django-imagekit==6.1.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-robots==6.1
django-storages==1.14.6
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
Faker==40.11.1
feedgenerator==2.2.1
gunicorn==25.2.0
idna==3.11
inflection==0.5.1
itsdangerous==2.2.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
Markdown==3.10.2
packaging==26.0
pilkit==3.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
s3transfer==0.16.0
sentry-sdk==2.56.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
whitenoise==6.12.0
